refactor js to ESM
