+++ README.md
@@ -3,8 +3,8 @@
 
 4주차: 빌드와 데이터 통신
 
-- [ ] 지난주 구현 사항에서 부족한 부분 보완하기
-* ESM 방식으로 JS 파일 분리
+- [x] 지난주 구현 사항에서 부족한 부분 보완하기
+* ~~ESM 방식으로 JS 파일 분리~~
 * ~~페이지네이션 구현~~
 * ~~슬라이드에서 넘어가는 카드 인자로 받는 기능 구현~~
