+++ index.html
@@ -3,7 +3,7 @@
     <head>
         <link rel="stylesheet" href="css/style.css">
         <title>Hyungyu Jung's Static Netflix Clone Webpage</title>
-        <script src="app.js" defer></script>
+        <script type="module" src="js/main.js" defer></script>
     </head>
     <body class="page">
         <header class="header">
