+++ js/main.js
@@ -1,0 +1,5 @@
+import { initSliders } from './slider.js';
+
+document.addEventListener('DOMContentLoaded', () => {
+    initSliders();
+});
